Updated build configuration.
Configuration paramaters such as framework search path, header search
path, build product location, and several other folders are now passed
to the Build Framework and Build Extensions scripts instead of being
defined in their configuration files. This is done so that only the
main Textual configuration needs to be updated for the passing of this
information.

Additionally, the actual configuration of each target has been updated
to share the same values. This includes the main Textual executable,
the frameworks, and the extensions.

Build results for everything is now stored in a .tmp folder in the root
directory. The Textual executable result is still located in the
"BuildResults" folder.

+++ Frameworks/Blowfish Encryption/Source/Classes/BlowfishBase.m
@@ -67,13 +67,13 @@ static const signed char fish_unbase64[256] = {
 	const char *message	= [rawInput	 cStringUsingEncoding:dataEncoding];
 	const char *secrkey	= [secretKey cStringUsingEncoding:dataEncoding];
 
-	size_t keylen = secretKey.length;
-	size_t msglen = rawInput.length;
+	NSInteger keylen = secretKey.length;
+	NSInteger msglen = rawInput.length;
 
 	/* =============================================== */
 
 	BF_KEY bfkey;
-    BF_set_key(&bfkey, keylen, (const unsigned char *)secrkey);
+    BF_set_key(&bfkey, (int)keylen, (const unsigned char *)secrkey);
 
 	NSInteger mallocSize = msglen;
 
@@ -157,13 +157,13 @@ static const signed char fish_unbase64[256] = {
 	const char *message	= [rawInput	 cStringUsingEncoding:dataEncoding];
 	const char *secrkey	= [secretKey cStringUsingEncoding:dataEncoding];
 
-	size_t keylen = secretKey.length;
-	size_t msglen = rawInput.length;
+	NSInteger keylen = secretKey.length;
+	NSInteger msglen = rawInput.length;
 
 	/* =============================================== */
 
 	BF_KEY bfkey;
-    BF_set_key(&bfkey, keylen, (const unsigned char *)secrkey);
+    BF_set_key(&bfkey, (int)keylen, (const unsigned char *)secrkey);
 
     char *decrypted = malloc((msglen + 1));
     char *end = decrypted;
